requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
cramjam==2.8.3
    # via fastparquet
fastparquet==2024.2.0
    # via -r requirements.in
fsspec==2024.3.1
    # via fastparquet
greenlet==3.0.3
    # via sqlalchemy
numpy==1.26.4
    # via
    #   fastparquet
    #   pandas
    #   pyarrow
packaging==24.0
    # via fastparquet
pandas==2.2.2
    # via
    #   -r requirements.in
    #   fastparquet
psycopg2-binary==2.9.9
    # via -r requirements.in
pyarrow==16.1.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.1
    # via pandas
six==1.16.0
    # via python-dateutil
sqlalchemy==2.0.30
    # via -r requirements.in
typing-extensions==4.11.0
    # via sqlalchemy
tzdata==2024.1
    # via pandas
